Revert "Disable the legacy gtest."
Apparently there are dozens of other builds using this that aren't
using the Android build system. I'm not going to go fix all of them,
so just put all the source back.

This project has been removed from the master manifest, so nothing
here should be colliding with the real one in external/googletest.

This reverts commit 0af8635d8866bdfdc72b1070f041ada1ac136e3f.

Bug: http://b/16574165
Change-Id: Ie07a8156281a1dab9885eef64788fe362d4e3c36